Move connection.py to __init__.py
