+++ plugwise/connections/serial.py
@@ -3,7 +3,7 @@ import logging
 
 import serial
 
-from plugwise.connections.connection import StickConnection
+from plugwise.connections import StickConnection
 from plugwise.constants import BAUD_RATE, BYTE_SIZE, STOPBITS
 from plugwise.exceptions import PortError
